requirements.txt
selenium==3.141.0
openai==0.10.3
python-dotenv==0.19.2

#Use pip install -r requirements.txt to Install this MODULE